Example of using partials in hugo
Partials are used to store the shared piece of code which repeats in more than one templates.
Partials are kept in themes/<themename>/layouts/partials directory
Partials can be used to override the themes presentation.
Use partials as much as possible.

Here we simply created header and footer and used it in index.html, single.html and list.html.
This way we don't have to repeat the contents.