Fixing the date shown on About page
We will create a new section template to fix this issue.
Here we create a section template for posts which will override the "_default" single page.
We keep the Date metadata display in here. This template will be used to render the pages under "post" section.

We removed the display of date metadata in the "_default" single page template which will be used
to render about.md.